+++ public/components/ui_action_context_menu.tsx
@@ -16,16 +16,21 @@ import { i18n } from '@osd/i18n';
 
 import { BehaviorSubject } from 'rxjs';
 import { useObservable } from 'react-use';
+import { HttpSetup } from '../../../../src/core/public';
 import { buildContextMenuForActions } from '../../../../src/plugins/ui_actions/public';
 import { AI_ASSISTANT_QUERY_EDITOR_TRIGGER } from '../ui_triggers';
 import { getUiActions } from '../services';
 import { DataPublicPluginSetup } from '../../../../src/plugins/data/public';
-
-interface Props {
+import { DATA2SUMMARY_AGENT_CONFIG_ID } from '../../common/constants/llm';
+import { AssistantServiceStart } from '../services/assistant_service';
+export interface Props {
   data: DataPublicPluginSetup;
   isQuerySummaryCollapsed$: BehaviorSubject<boolean>;
   resultSummaryEnabled$: BehaviorSubject<boolean>;
+  isSummaryAgentAvailable$: BehaviorSubject<boolean>;
+  httpSetup: HttpSetup;
   label?: string;
+  assistantServiceStart: AssistantServiceStart;
 }
 
 export const ActionContextMenu = (props: Props) => {
@@ -39,6 +44,26 @@ export const ActionContextMenu = (props: Props) => {
   });
   const resultSummaryEnabled = useObservable(props.resultSummaryEnabled$, false);
   const isQuerySummaryCollapsed = useObservable(props.isQuerySummaryCollapsed$, false);
+  const isSummaryAgentAvailable = useObservable(props.isSummaryAgentAvailable$, false);
+  const shouldShowSummarizationAction = resultSummaryEnabled && isSummaryAgentAvailable;
+
+  useEffect(() => {
+    if (!resultSummaryEnabled) return;
+    props.isSummaryAgentAvailable$.next(false);
+    const fetchSummaryAgent = async () => {
+      try {
+        const summaryAgentStatus = await props.assistantServiceStart.client.agentConfigExists(
+          DATA2SUMMARY_AGENT_CONFIG_ID,
+          { dataSourceId: props.data.query.queryString.getQuery().dataset?.dataSource?.id }
+        );
+        props.isSummaryAgentAvailable$.next(!!summaryAgentStatus.exists);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+
+    fetchSummaryAgent();
+  }, [props.data.query.queryString.getQuery()?.dataset?.dataSource?.id, resultSummaryEnabled]);
 
   useEffect(() => {
     const subscription = props.data.query.queryString.getUpdates$().subscribe((query) => {
@@ -75,13 +100,14 @@ export const ActionContextMenu = (props: Props) => {
     [actionContext.datasetId, actionContext.datasetType, actionContext.dataSourceId]
   );
 
-  // The action button should be not displayed when there is no action and result summary disabled.
-  if (actionsRef.current.length === 0 && !resultSummaryEnabled) {
+  // The action button should be not displayed when there is no action and result summary disabled or there is no data2Summary agent
+  if (!shouldShowSummarizationAction && actionsRef.current.length === 0) {
     return null;
   }
 
-  // The action button should be disabled when context menu has no item and result summary disabled.
-  const actionDisabled = (panels.value?.[0]?.items ?? []).length === 0 && !resultSummaryEnabled;
+  // The action button should be disabled when context menu has no item or result summary disabled or or no data2Summary agent is available
+  const actionDisabled =
+    !shouldShowSummarizationAction && (panels.value?.[0]?.items ?? []).length === 0;
 
   return (
     <EuiPopover
@@ -109,7 +135,7 @@ export const ActionContextMenu = (props: Props) => {
       closePopover={() => setOpen(false)}
     >
       <EuiContextMenu size="s" initialPanelId={'mainMenu'} panels={panels.value} />
-      {resultSummaryEnabled && (
+      {shouldShowSummarizationAction && (
         <EuiPopoverFooter paddingSize="s">
           <EuiSwitch
             label={i18n.translate('queryEnhancements.queryAssist.summary.switch.label', {